Feat: opensearch 서비스 분리
