--- src/main/java/site/hesil/latteve_spring/domains/search/service/SearchIndexService.java
@@ -1,15 +1,0 @@
-package site.hesil.latteve_spring.domains.search.service;
-
-/**
- * packageName    : site.hesil.latteve_spring.domains.search.service
- * fileName       : SearchIndexService
- * author         : Heeseon
- * date           : 2024-08-30
- * description    :
- * ===========================================================
- * DATE              AUTHOR             NOTE
- * -----------------------------------------------------------
- * 2024-08-30        Heeseon       최초 생성
- */
-public class SearchIndexService {
-}
