+++ src/main/java/site/hesil/latteve_spring/domains/search/service/SearchService.java
@@ -60,151 +60,10 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class SearchService {
-    private final MemberRepository memberRepository;
-    private final MemberStackRepository memberStackRepository;
-    private final ProjectRepository projectRepository;
+
     private final OpenSearchClient openSearchClient;
-    private final ProjectStackRepository projectStackRepository;
-    private final ProjectMemberRepository projectMemberRepository;
-    private final RecruitmentRepository recruitmentRepository;
-    private final TechStackRepository techStackRepository;
-    private final MemberJobRepository memberJobRepository;
-    private final JobRepository jobRepository;
-    private final ProjectLikeRepository projectLikeRepository;
-    private final SortHandlerMethodArgumentResolverCustomizer sortCustomizer;
-
-    // 나중에 삭제 -> 데이터 동기화
-    @PostConstruct
-    public void init() throws IOException {
-        createOrRecreateIndexWithMapping("projects"); // projects 인덱스 생성 또는 재생성
-        createOrRecreateIndexWithMapping("members");  // members 인덱스 생성 또는 재생성
-        indexProjectsToOpenSearch();  // 프로젝트 데이터 인덱싱
-        indexMembersToOpenSearch(); // 멤버 데이터 인덱싱
-    }
-    private void createOrRecreateIndexWithMapping(String indexName) throws IOException {
-        // 인덱스 존재 여부 확인
-        boolean exists = openSearchClient.indices().exists(e -> e.index(indexName)).value();
-
-        if (exists) {
-            // 인덱스가 존재할 경우 삭제
-            openSearchClient.indices().delete(new DeleteIndexRequest.Builder().index(indexName).build());
-        }
-
-        // 인덱스별 techStack 매핑 설정
-        TypeMapping mapping = createMapping(indexName);
-
-        // 인덱스 생성 요청
-        CreateIndexRequest createIndexRequest = new CreateIndexRequest.Builder()
-                .index(indexName)
-                .mappings(mapping)
-                .build();
-
-        // 인덱스 생성
-        CreateIndexResponse createIndexResponse = openSearchClient.indices().create(createIndexRequest);
-
-        if (!createIndexResponse.acknowledged()) {
-            throw new RuntimeException("Failed to create OpenSearch index with mappings.");
-        }
-    }
-
-
-    private TypeMapping createMapping(String indexName) {
-        String filedName;
-        Property techStackProperty = Property.of(b -> b
-                .nested(o -> o
-                        .properties("key", Property.of(pb -> pb.text(t -> t)))
-                        .properties("value", Property.of(pb -> pb.text(t -> t)))
-                )
-        );
-        Map<String, Property> properties = new HashMap<>();
-        if("projects".equals(indexName)) {
-            properties.put("projectTechStack", techStackProperty);
-            // project의 name 필드에 keyword 서브 필드 추가
-            properties.put("name", createTextFieldWithKeyword());
-        }else if("members".equals(indexName)){
-            properties.put("techStack", techStackProperty);
-            // memberNickname 필드에 keyword 서브 필드 추가
-            properties.put("memberNickname", createTextFieldWithKeyword());
-        }else {
-            throw new IllegalArgumentException("Unknown index name: " + indexName);
-        }
-        // 매핑 생성
-        return TypeMapping.of(mb -> mb
-                .properties(properties)
-        );
-    }
-
-    // text 타입인 필드에 keyword 서브 필드 추가
-    private Property createTextFieldWithKeyword(){
-        return Property.of(b-> b.text(t -> t.fields("keyword", Property.of(pb->pb.keyword(k -> k )))));
-    }
 
     /** project search */
-    @Transactional
-    public void indexProjectsToOpenSearch() throws IOException {
-        List<Project> projects = projectRepository.findAll();
-        //Project -> Index
-        for(Project project : projects) {
-            // 프로젝트에 연관된 기술 스택  정보 가져옴
-            List<ProjectStack> projectTechStacks = projectStackRepository.findAllByProject_ProjectId(project.getProjectId());
-
-            // techStack list로 저장
-            List<ProjectDocumentReq.TechStack> techStackList= new ArrayList<>();
-            for (ProjectStack projectStack : projectTechStacks) {
-                Optional<TechStack> techStackOpt = techStackRepository.findById(projectStack.getTechStack().getTechStackId());
-                if (techStackOpt.isPresent()) {
-                    TechStack techStack = techStackOpt.get();
-                    String name = techStack.getName();
-                    String imgUrl = techStack.getImgUrl() != null ? techStack.getImgUrl() :
-                            (projectStack.getCustomStack() != null ? projectStack.getCustomStack() : name);
-                    techStackList.add(new ProjectDocumentReq.TechStack(name, imgUrl));
-                }
-            }
-
-            String statusToString = convertStatusToString(project.getStatus());
-
-
-            // 좋아요 수
-            Long cntLike = projectLikeRepository.countProjectLikeByProject_ProjectId(project.getProjectId());
-
-            // 프로젝트에 필요한 인원
-            Integer requiredMemberCount = recruitmentRepository.findMemberCountByProject_ProjectId(project.getProjectId());
-            // 프로젝트에 지원한 인원
-            Integer currentMemberCount = projectMemberRepository.findMemberCountByProject_ProjectId(project.getProjectId());
-
-            // ProjectDocumentReq 생성
-            ProjectDocumentReq projectDocumentReq = ProjectDocumentReq.builder()
-                    .projectId(project.getProjectId())
-                    .name(project.getName())
-                    .imgUrl(project.getImgUrl())
-                    .duration(project.getDuration())
-                    .projectTechStack(techStackList)
-                    .teamCnt(requiredMemberCount)
-                    .currentCnt(currentMemberCount)
-                    .cntLike(cntLike)
-                    .status(statusToString)
-                    .createdAt( formatLocalDateTime(project.getCreatedAt()))
-                    .build();
-
-            // Elasticsearch에 인덱싱
-            IndexRequest<ProjectDocumentReq> indexRequest = new IndexRequest.Builder<ProjectDocumentReq>()
-                    .index("projects")
-                    .id(project.getProjectId().toString())
-                    .document(projectDocumentReq)
-                    .build();
-            openSearchClient.index(indexRequest);
-        }
-    }
-
-    private String convertStatusToString(int status) {
-        switch (status) {
-            case 0: return "모집중";
-            case 1: return "진행중";
-            case 2: return "종료";
-            default: return "Unknown";
-        }
-    }
-
     public List<ProjectDocumentReq> searchProjectsByKeyword(String keyword, String status, String sort) throws IOException {
 
         // keyword로 검색
@@ -265,80 +124,7 @@ public class SearchService {
     }
 
     /**Member search */
-    @Transactional
-    public void indexMembersToOpenSearch() throws IOException {
-        List<Member> members = memberRepository.findAll();
-        // Member -> Index
-        for (Member member : members) {
-            // 멤버에 연관된 기술 스택 정보 가져옴
-            List<MemberStack> memberStacks = memberStackRepository.findAllByMember_MemberId(member.getMemberId());
-            // techStack list로 저장
-            List<MemberDocumentReq.TechStack> techStackList = new ArrayList<>();
-            for (MemberStack memberStack : memberStacks) {
-                Optional<TechStack> techStackOpt = techStackRepository.findById(memberStack.getTechStack().getTechStackId());
-                if (techStackOpt.isPresent()) {
-                    TechStack techStack = techStackOpt.get();
-                    String name = techStack.getName();
-                    String imgUrl = techStack.getImgUrl() != null ? techStack.getImgUrl() : (memberStack.getCustomStack() != null ? memberStack.getCustomStack() : name);
 
-                    // TechStack 객체를 리스트에 추가
-                    techStackList.add(new MemberDocumentReq.TechStack(name, imgUrl));
-                }
-            }
-            // 멤버의 직무 정보 가져옴
-            List<MemberJob> memberJobs = memberJobRepository.findAllByMember_MemberId(member.getMemberId());
-
-            List<String> jobList = new ArrayList<>();
-            // 멤버의 직무 이름 list로 저장
-            for(MemberJob memberJob : memberJobs) {
-                Optional<Job> jobOpt = jobRepository.findById(memberJob.getJob().getJobId());
-                jobOpt.ifPresent(job -> jobList.add(job.getName()));
-
-            }
-
-            // Member가 참여한 프로젝트 개수
-            int ongoingProjectCount = projectRepository.countProjectsByMemberIdAndStatus(member.getMemberId(), 1);
-            int completedProjectCount = projectRepository.countProjectsByMemberIdAndStatus(member.getMemberId(), 2);
-
-            // careerSortValue 계산
-            int careerSortValue = calculateCareerSortValue(Collections.singletonList(member.getCareer()));
-
-            // MemberDocumentReq 생성
-            MemberDocumentReq memberDocumentReq = MemberDocumentReq.builder()
-                    .memberId(member.getMemberId())
-                    .memberNickname(member.getNickname())
-                    .memberImg(member.getImgUrl())
-                    .memberGithub(member.getGithub())
-                    .techStack(techStackList)  // TechStack 리스트 전달
-                    .ongoingProjectCount(ongoingProjectCount)
-                    .completedProjectCount(completedProjectCount)
-                    .memberJob(jobList)
-                    .career(member.getCareer())
-                    .careerSortValue(careerSortValue)
-                    .createdAt(formatLocalDateTime(member.getCreatedAt()))
-                    .build();
-
-            // Elasticsearch에 인덱싱
-            IndexRequest<MemberDocumentReq> indexRequest = new IndexRequest.Builder<MemberDocumentReq>()
-                    .index("members")
-                    .id(member.getMemberId().toString())
-                    .document(memberDocumentReq)
-                    .build();
-            openSearchClient.index(indexRequest);
-        }
-    }
-
-    private int calculateCareerSortValue(List<String> career) {
-        // 경력 직무에 따라 정수 값 설정 (예시: 시니어 3, 주니어 2, 신입 1)
-        if (career.contains("시니어")) {
-            return 3;
-        } else if (career.contains("주니어")) {
-            return 2;
-        } else if (career.contains("신입")) {
-            return 1;
-        }
-        return 0; // 기타 또는 알 수 없는 경우
-    }
     public List<MemberDocumentReq> searchMembersByKeyword(String keyword , String sort) throws IOException {
 
         BoolQuery boolQuery = BoolQuery.of(b -> {
@@ -443,8 +229,5 @@ public class SearchService {
     }
 
 
-    public String formatLocalDateTime(LocalDateTime localDateTime) {
-        DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-        return localDateTime.format(formatter);
-    }
+
 }
